split parser.py to two file parser_link.py and parser_info.py
